FIR: extract common logic from KotlinNameSuggester and share
