+++ idea/idea-fir/src/org/jetbrains/kotlin/idea/core/FirKotlinNameSuggester.kt
@@ -1,0 +1,8 @@
+/*
+ * Copyright 2010-2021 JetBrains s.r.o. and Kotlin Programming Language contributors.
+ * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
+ */
+
+package org.jetbrains.kotlin.idea.core
+
+object FirKotlinNameSuggester : AbstractNameSuggester()
